Remove manually sorted test data from treemap tests
Since the retrieval test already checks that correct values are
retrieved with corresponding keys, it should be enough for the
iterator tests to check that the iterator returns nodes in the
correct order of keys.

That can be done by checking that each key returned by the iterator
is greater than the previous one, without having a sorted version
of the input test data hard-coded or otherwise separately sorted.